Factored out the RDF::Enumerable#count and #empty? methods into an RDF::Countable mixin.
Now that we are as of RDF.rb 0.2.0 generally returning enumerators instead of arrays, we've also inadvertently introduced a bit of a usability problem: enumerables and enumerators don't support the #empty? or #size methods.

This new mixin defines #empty? and #count (and the latter's alias #size) in terms of #each. It can be mixed into any enumerator return value that reasonably enough should also be countable.