Factored out the append-only bits of RDF::Mutable into RDF::Writable.
In particular, the #<< and #insert methods are now defined in RDF::Writable instead of RDF::Mutable (though the latter includes everything from the former, of course).

This change benefits stream-oriented classes, such as RDF::Writer, which will want to mix in RDF::Writable but not the rest of RDF::Mutable.