Reverted changes to RDF::Literal::{Decimal,Double}#to_s.
The RDF::Literal#to_s method does not need to (and in general perhaps should not) perform canonicalization. According to the principle of least surprise, we don't want e.g. RDF::Literal.new(0.0).to_s to return "0.0E0", we want it to return "0.0". We should also ensure that #to_s remains a relatively cheap operation.