Replaced << in initialize with unshift
Should be O(n) instead of O(n^2/2). Also confirms behavior of unshift,
and led to recognition that unshift needs a second optional arg: the subject of
the new element.

Note: there's a slight divergence in behavior: creating an new empty
list with an explicit subject won't "unshift" into that node.