Make `Transaction#execute` more atomic
Executing a transaction individually called `#delete` or `#insert` for
each statement in the changeset. This prevents using the features of
`Repositories` with a more efficient implementation of
`#insert/delete_statements`.

The patch fixes this (partially) by calling `#delete` and `#insert` only
once each for each execution, making executions faster for some
repositories and closer to being properly atomic.

A caveat, of course, is that there are still two upstream requests
required for each "transaction". We may regard that as okay, with
@bendiken's planned 2.0 changes to transaction, and seeing this
interface as a changeset.

Another option is to implement `#delete_insert(deletes, inserts)` in
`RDF::Mutable`; the base implementation could simply be to call
`#delete` and `#insert` in turn, but `Repository` specializations could
handle them as a single request. Thoughts?