Fix a typo (`Enummerable`) in `RDF::Mutable`
This was introduced in 22d6fea51, leading up to the 1.99 release.

This conditional is apparently untested; apparently we never try to
`#update` on a single statement throughout the suite. I'm working on a
test to catch this.

I believe we should merge this immediately, since it is a simple typo
fix. Any issues revealed in adding `#update` testing can be dealt with
as an issue unrelated to the typo.