Transaction#rollback & snapshot isolation interface
Defines transaction rollback requirements and implements them for the
base class.

Adds support for passing a isolated snapshot of the repository. The
default is to use an unisolated repository for query; but this supports
using any Queryable as a snapshot.

Further documentation for adding transaction support to a Repository
using the base class.