Faster `#has_graph?` and `#graph_names?`
The base repository implementation had removed `#graph_names` and
`#has_graph?`, falling back on the versions of those methods in
Enumerable. This avoids the need to touch each statement by checking the
keys.

This is the same implementation as in the older (Hash based) version of
Repository, slightly modified for Hamster.