Performance tweaks in Repository::Implementation
There are two minor performance tweaks in `#query_pattern` here, with a
measurable effect:

First, we avoid checking `#has_key?` if the sought value is a
variable. The nil/type check should be cheaper than the key search.

Second, we remove a double check for subjects and predicates.

We also use `#has_key?` in `#has_graph?`, rather than
`#keys.include?`. I'm uncertain whether this has a real benefit, but it
pushes responsibility for an efficient key search to `Hamster`. We might
be avoiding touching each key by making this change.