Use `#any?` in `Repository::Implementation#has_statement?` instead of `#include?`.
`Repository::Implementation#has_statement?` is broken compared to the default Hash implementation because objects are maintained with a `Hamster::Set` rather than an `Array`. There is a different between `(Hamster::)Set`#include? and `Array#include?` in that `Array#include?` uses == for comparison, and Set looks for the same object (similar to a Hash key). Query depends on the use of == to do appropriate L2V mapping.