Fixup #isolation_level handling
Certain aspects of `#isolation_level` don't apply to snapshots, but to
snapshot + transaction. This peels the two apart somewhat.

`Dataset` & `Repository` handle communicating the isolation level for
`#snapshot`.

The base `RDF::Transaction` defaults to `:read_committed`, but adheres
to the snapshot's semantics if present. Generally, this will mean
`:repeatable_read` or worse.

The default Repository's `SerializedTransaction` assumes a snapshot, and
is always `:serializable`.