Add graph scopes to queryable transactions
This changes the Transaction semantics to behave like Graph. Queryable
is somewhat opinionated that inserting triples to a queryable should
result in those triples being accessible through query. In light of
this, the answer to #278 seems to be that graph scope must be
symmetrical for read and write operations. That interpretation is
implemented here.