Add support for `:literal_equality`
This allows Enumerables to decline to preserve term-inequality for
literals that are value-equal.

For example: `"01"^^xsd:integer` and `"1"^^xsd:integer` have equal
values, but are not equal per RDF term-equality. An enumerable
supporting `:literal_equality` must preserve this distinction across
reads; one that does not may canonicalize the lexical representation of
its literals at its own discretion.

Note that an `Enumerable` need not be `Mutable` to perform this
canonicalization. For example, a data store may have an "optimize"
operation which performs the canonicalization independent of RDF.rb's
interactions.