Upgrade Hamster to 3.0.0
This upgrades the Hamster dependency to the recently released 3.0.0. The
breaking change in that release is in the behavior of
`Hamster::Hash#values_at`, which is unused.

It also adds superset/subset operators for `Hamster::Hash`, which may be
of interest to us in handling smart Repository merges.

See the diff, including Changelog, at: https://github.com/hamstergem/hamster/compare/v2.0.0...v3.0.0