Implement `Transaction#mutated?`
`#mutated?` must return `true` when the transaction would change the
repository if executed against it's state at transaction start time.
We allow implementations to return `true` in other cases, as well; there
is no guarantee that a `false` result will be returned in any specific
circumstances (or ever).

For the base transaction, we simply check whether `@changes` is
empty. This can give false positives, but avoids a worst-case linear
check.

For subclasses with no custom implementation, we raise a
`NotImplementedError`.

For the `SerializableTransaction` included in the default
`RepositoryImplementation`, we use `#equal?` on the `Hamster::Hash`
instances. This will always give a correct response for the actual
implementation, moreover, it is guaranteed not to give false negatives
even if a user messes with the snapshot using `#send`.