Update #start_with? to match String#start_with signature
Faraday recently started expecting a string-like #start_with? signature,
where you can pass multiple strings to test against and it returns true
if any of them passes. We were passing an RDF Value to Faraday since the
API was mostly consistent.

It seems like the intent is to have the same API as String#start_with?,
so this replicates it.